+++ deprecated/users/auth.ts
